+++ dubbo-demo/dubbo-demo-spring-boot/dubbo-demo-spring-boot-provider/src/main/resources/application.properties
@@ -1,0 +1,29 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+spring.application.name=dubbo-springboot-demo-provider
+
+dubbo.application.name=${spring.application.name}
+dubbo.protocol.name=dubbo
+dubbo.protocol.port=-1
+dubbo.registry.address=zookeeper://127.0.0.1:2181
+dubbo.registry.id=zk-registry
+dubbo.config-center.address=zookeeper://127.0.0.1:2181
+dubbo.metadata-report.address=zookeeper://127.0.0.1:2181
+dubbo.metrics.protocol=prometheus
+dubbo.metrics.enable-jvm=true
+dubbo.scan.base-packages=org.apache.dubbo.samples.provider.service
+
