+++ dubbo-spring-boot/dubbo-spring-boot-starters/observability/dubbo-tracing-otel-otlp-spring-boot-starter/pom.xml
@@ -31,7 +31,7 @@
   <dependencies>
     <dependency>
       <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-observability-spring-boot-autoconfigure</artifactId>
+      <artifactId>dubbo-spring-boot-autoconfigure</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
