+++ dubbo-spring-boot/dubbo-spring-boot-starters/observability/pom.xml
@@ -28,59 +28,10 @@
   <packaging>pom</packaging>
 
   <modules>
-    <module>autoconfigure</module>
     <module>dubbo-tracing-otel-zipkin-spring-boot-starter</module>
     <module>dubbo-tracing-brave-zipkin-spring-boot-starter</module>
     <module>dubbo-observability-spring-boot-starter</module>
     <module>dubbo-tracing-otel-otlp-spring-boot-starter</module>
   </modules>
-
-  <properties>
-    <micrometer.version>1.13.3</micrometer.version>
-    <micrometer-tracing.version>1.3.3</micrometer-tracing.version>
-    <opentelemetry.version>1.42.0</opentelemetry.version>
-    <zipkin-reporter.version>3.4.0</zipkin-reporter.version>
-    <prometheus-client.version>0.16.0</prometheus-client.version>
-  </properties>
-
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>io.micrometer</groupId>
-        <artifactId>micrometer-bom</artifactId>
-        <version>${micrometer.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
-        <groupId>io.micrometer</groupId>
-        <artifactId>micrometer-tracing-bom</artifactId>
-        <version>${micrometer-tracing.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
-        <groupId>io.opentelemetry</groupId>
-        <artifactId>opentelemetry-bom</artifactId>
-        <version>${opentelemetry.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
-        <groupId>io.zipkin.reporter2</groupId>
-        <artifactId>zipkin-reporter-bom</artifactId>
-        <version>${zipkin-reporter.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-      <dependency>
-        <groupId>io.prometheus</groupId>
-        <artifactId>simpleclient_bom</artifactId>
-        <version>${prometheus-client.version}</version>
-        <type>pom</type>
-        <scope>import</scope>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
 
 </project>
