+++ dubbo-test/dubbo-dependencies-all/pom.xml
@@ -390,11 +390,6 @@
 
     <dependency>
       <groupId>org.apache.dubbo</groupId>
-      <artifactId>dubbo-observability-spring-boot-autoconfigure</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.dubbo</groupId>
       <artifactId>dubbo-tracing-otel-zipkin-spring-boot-starter</artifactId>
       <version>${project.version}</version>
     </dependency>
